A punto de entregar

+++ src/main/java/Backend/Services/ServerContract.java
@@ -14,8 +14,8 @@ public interface ServerContract extends RemoteTabla {
     public final static int TYPE_MORE_THAN = 1;
     public final static int TYPE_LESS_THAN = 2;
 
-    void register(ClientContract callbackClientObject, int i, int type, int cuantity) throws RemoteException;
-    void unregister(ClientContract callbackClientObject, int i, int type, int cuantity) throws RemoteException;
+    void register(ClientContract callbackClientObject, int i, int type, float cuantity) throws RemoteException;
+    void unregister(ClientContract callbackClientObject, int i, int type, float cuantity) throws RemoteException;
     void unregister(ClientContract callbackClientObject) throws RemoteException;
     void sendnotify() throws RemoteException;
     ArrayList<String> getElements() throws RemoteException;
